move test stack
